TMS_ETL/requirements.txt
pandas==1.5.3
pyodbc==4.0.39
python-dotenv==1.0.0
requests==2.28.2
openpyxl==3.1.2
